Fix compiler errors

+++ src/race/race_manager.hpp
@@ -69,7 +69,7 @@ static const std::string IDENT_CUTSCENE ("CUTSCENE"        );
  *     manager stores the GP information, but World queries only track
  *     and number of laps, so in case of GP this information is taken from
  *     the GrandPrix object), and local player information (number of local
- *     players, and selected karts). 
+ *     players, and selected karts).
  *     Information about player karts (which player selected which kart,
  *     player ids) is stored in a RemoteKartInfo structure and used later
  *     to initialise the KartStatus array (startNew()). The KartStatus array
@@ -524,7 +524,7 @@ public:
     {
         return m_num_local_players;
     }   // getNumLocalPlayers
-    
+
     // ----------------------------------------------------------------------------------------
     /** Returns true if the split screen display leaves an empty space that
      *  can be used to display the minimap.
@@ -544,7 +544,7 @@ public:
     // ----------------------------------------------------------------------------------------
     unsigned int getNumberOfAIKarts() const
     {
-        return (unsigned int)m_ai_kart_list.size(); 
+        return (unsigned int)m_ai_kart_list.size();
     }   // getNumberOfAIKarts
     // ----------------------------------------------------------------------------------------
     unsigned int getNumberOfRedAIKarts() const { return m_num_red_ai; }
@@ -574,9 +574,9 @@ public:
         }
     }
     // ----------------------------------------------------------------------------------------
-    unsigned int getNumPlayers() const 
+    unsigned int getNumPlayers() const
     {
-        return (unsigned int) m_player_karts.size(); 
+        return (unsigned int) m_player_karts.size();
     }   // getNumPlayers
     // ----------------------------------------------------------------------------------------
     /** \brief Returns the number lf laps.
@@ -754,7 +754,7 @@ public:
 
     // ----------------------------------------------------------------------------------------
     bool isCTFMode() const       { return m_minor_mode == MINOR_MODE_CAPTURE_THE_FLAG; }
- 
+
     // ----------------------------------------------------------------------------------------
     bool isEggHuntMode() const   { return m_minor_mode == MINOR_MODE_EASTER_EGG; }
 
@@ -919,7 +919,7 @@ public:
     }
     // ----------------------------------------------------------------------------------------
     [[nodiscard]]
-    std::optional<size_t> getCurrentRaceId() const;
+    std::optional<uint64_t> getCurrentRaceId() const;
     // ----------------------------------------------------------------------------------------
     [[nodiscard]]
     RestApi::Server& getServer() noexcept { return *m_server; }
