+++ src/rest-api/RaceObserver.hpp
@@ -13,9 +13,9 @@ public:
     static std::unique_ptr<RaceResultLoader> create();
 
 public:
-    [[nodiscard]] virtual std::optional<size_t> getLatestId() const = 0;
-    [[nodiscard]] virtual std::string get(size_t id) const = 0;
-    virtual void store(size_t id, const std::string& result) = 0;
+    [[nodiscard]] virtual std::optional<uint64_t> getLatestId() const = 0;
+    [[nodiscard]] virtual std::string get(uint64_t id) const = 0;
+    virtual void store(uint64_t id, const std::string& result) = 0;
 };
 
 class RaceObserver
@@ -25,13 +25,13 @@ public:
     void start();
     void update();
     void stop();
-    [[nodiscard]] std::optional<size_t> getCurrentRaceId() const;
+    [[nodiscard]] std::optional<uint64_t> getCurrentRaceId() const;
 
 private:
     RaceResultLoader& loader_;
     std::function<std::string()> getCurrentState_;
     bool active_;
-    std::optional<size_t> id_;
+    std::optional<uint64_t> id_;
 };
 
 }
