+++ src/test/rest-api/MockRaceResultLoader.hpp
@@ -5,7 +5,7 @@
 class MockRaceResultLoader : public RestApi::RaceResultLoader
 {
 public:
-    MOCK_METHOD(std::optional<size_t>, getLatestId, (), (const, override));
-    MOCK_METHOD(std::string, get, (size_t), (const, override));
-    MOCK_METHOD(void, store, (size_t, const std::string&), (override));
+    MOCK_METHOD(std::optional<uint64_t>, getLatestId, (), (const, override));
+    MOCK_METHOD(std::string, get, (uint64_t), (const, override));
+    MOCK_METHOD(void, store, (uint64_t, const std::string&), (override));
 };
